requirements.txt
mss==4.0.3
numpy==1.16.4
opencv-python==4.1.0.25
pypiwin32
